Move argument parser to own module
