+++ src/lib.rs
@@ -144,6 +144,7 @@ extern crate proc_macro;
 #[macro_use]
 mod quote;
 
+mod args;
 mod error;
 mod iter;
